backend/requirements.txt
fastapi>=0.120.0
uvicorn[standard]>=0.38.0
sqlalchemy[postgresql]>=2.0.44
alembic>=1.17.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
recipe-scrapers>=14.54.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
pdf2image>=1.16.3
selenium>=4.16.0
webdriver-manager>=4.0.1
playwright>=1.40.0
PyMuPDF>=1.24.0
pdfminer.six>=20221105
paddleocr>=2.7.0
anthropic>=0.31.0
google-generativeai>=0.3.0
Pillow>=10.0.0
